make a minor change because minecraft is Bad.
I was GOING to add an enchantment but it turns out that healing the player by a precise amount is not something minecraft likes doing

dear mc devs: why, pray tell, is damaging entities by negative amounts unsupported? is it just because dealing negative amounts of damage is obviously a nonsensical notion that could never ever be desirable or useful?